--- a/modal/_container_entrypoint.py
+++ b/modal/_container_entrypoint.py
@@ -618,13 +618,24 @@ async def call_function_async(
                         await function_io_manager._queue_put.aio(generator_queue, value)
                         item_count += 1
 
-                    await function_io_manager._queue_put.aio(
-                        generator_queue, _FunctionIOManager._GENERATOR_STOP_SENTINEL
-                    )
-                    await generator_output_task  # Wait to finish sending generator outputs.
                     message = api_pb2.GeneratorDone(items_total=item_count)
-                    await function_io_manager.push_output.aio(
-                        input_id, started_at, message, api_pb2.DATA_FORMAT_GENERATOR_DONE
+                    # Concurrently:
+                    #
+                    # 1. put stop sentinel to complete the generator outputs sending task
+                    # 2. wait to finish sending generator outputs
+                    # 3. send 'generator done' message with total item count
+                    #
+                    # Backend should not rely on a happens-before relationship between the
+                    # last item and the 'generator done' message, only that the latter message
+                    # will correctly report the total number of items to receive.
+                    await asyncio.gather(
+                        function_io_manager._queue_put.aio(
+                            generator_queue, _FunctionIOManager._GENERATOR_STOP_SENTINEL
+                        ),
+                        generator_output_task,
+                        function_io_manager.push_output.aio(
+                            input_id, started_at, message, api_pb2.DATA_FORMAT_GENERATOR_DONE
+                        ),
                     )
                 else:
                     if not inspect.iscoroutine(res) or inspect.isgenerator(res) or inspect.isasyncgen(res):