## Describe your changes


- In **local dev**, this reduces the high freq web endpoint's execution latency from ~9ms to ~3ms (66% latency reduction).  
- In **dev cluster**, this reduces the high freq web endpoint's execution latency from ~25ms to ~3ms (**88%** latency reduction). 
    - We can expect similar results in prod

### Repro

Take the highfreq `webhook.py` and deploy it with this change using `MODAL_SYNC_ENTRYPOINT=1`. Then hit it with `hyperfine` and observe the execution times in the app's logs. e.g.

`hyperfine -N --runs 100 'curl https://thundergolfer--highfreq-webhook-synthetic-monitor-webhook.modal-dev.run'`

### Backward/forward compatibility checks

Check these boxes or delete any item (or this section) if not relevant for this PR.

- [x] Client+Server: this change is compatible with old servers
- [x] Client forward compatibility: this change ensures client can accept data intended for later versions of itself

Note on protobuf: protobuf message changes in one place may have impact to
multiple entities (client, server, worker, database). See points above.


## Changelog

* Performance optimization to web endpoints, reducing 5-10ms of overhead.