@thundergolfer 2024-02-20
Thanks for review @ekzhang. Closing as this isn't correct behavior in the failure case.